Added paranoid check before rm -rf of cruft files from chalice deployment package in prune_chalice_package.sh.

+++ scripts/prune_chalice_package.sh
@@ -3,7 +3,7 @@
 # Script to remove (prune) unnecessary cruft from the specified Chalice zip file.
 # This was done when (2022-12-16) we reached the maximum (unzipped) size, 220476770 bytes,
 # of a Chalice package to deploy in AWS. To do this we delete all files/directories within
-# the zip file which look to be related to tests or example modules. The file is changed in place.
+# the zip file which look to be related to tests or example modules. The file is changed IN PLACE.
 # --------------------------------------------------------------------------------------------------
 
 if [ $# -ne 1 ]; then
@@ -32,9 +32,23 @@ mkdir ${TMP_CHALICE_PACKAGE_DIR}
 cd ${TMP_CHALICE_PACKAGE_DIR}
 unzip ${CHALICE_PACKAGE_ZIP_FILE} 2>&1 > $TMP_LOG_FILE
 
+# PLEASE BE CAREFUL HERE:
+# THIS below  ... the rm -rf ... looks understandably looks SCARY.
+# We ASSUME we are in (i.e. our current directory is) the TEMPORARY
+# directory (TMP_CHALICE_PACKAGE_DIR), which contains the contents
+# of our unzipped Chalice deployement package, via the cd above.
+# We will paranoia DOUBLE CHECK that this under /tmp first.
+#
+echo "Preparing to prune/delete cruft from chalice package."
+if [[ `pwd` != "/tmp/"* ]]; then
+    echo "ERROR: prune script has an error! In unxpected directory: `pwd`"
+    exit 1
+fi
+echo "Deleting cruft from chalice package."
 rm -rf `find . -name examples -type d`
 rm -rf `find . -name tests -type d`
 rm -rf `find . -name test -type d`
+echo "Done deleting cruft from chalice package."
 
 zip -r ${TMP_CHALICE_PACKAGE_FILE} . 2>&1 > $TMP_LOG_FILE
 rm -rf ${TMP_CHALICE_PACKAGE_DIR}
